[hotfix] rename the package name of mapper params from `mode` to `mapper`
