+++ flink-ml-parent/flink-ml-lib/src/main/java/org/apache/flink/ml/params/mapper/SISOFlatMapperParams.java
@@ -17,17 +17,17 @@
  * under the License.
  */
 
-package org.apache.flink.ml.params.mode;
+package org.apache.flink.ml.params.mapper;
 
-import org.apache.flink.ml.params.shared.colname.HasOutputCol;
+import org.apache.flink.ml.params.shared.colname.HasOutputColDefaultAsNull;
 import org.apache.flink.ml.params.shared.colname.HasReservedCols;
-import org.apache.flink.ml.params.shared.colname.HasSelectedCols;
+import org.apache.flink.ml.params.shared.colname.HasSelectedCol;
 
 /**
- * Parameters for MISOMapper.
+ * Params for SISOFlatMapper, include selectedCol, outputCol and reservedCol.
  */
-public interface MISOMapperParams<T> extends
-	HasSelectedCols<T>,
-	HasOutputCol<T>,
+public interface SISOFlatMapperParams<T> extends
+	HasSelectedCol<T>,
+	HasOutputColDefaultAsNull<T>,
 	HasReservedCols<T> {
 }
