+++ flink-ml-parent/flink-ml-lib/src/main/java/org/apache/flink/ml/params/mapper/SISOMapperParams.java
@@ -17,10 +17,10 @@
  * under the License.
  */
 
-package org.apache.flink.ml.params.mode;
+package org.apache.flink.ml.params.mapper;
 
 /**
- * Parameters for SISOModelMapper.
+ * Parameters for SISOMapper.
  */
-public interface SISOModelMapperParams<T> extends SISOMapperParams<T> {
+public interface SISOMapperParams<T> extends SISOFlatMapperParams<T> {
 }
